requirements.txt
streamlit
openai
streamlit-monaco
python-dotenv
pydantic
